extract TaskCard from TaskList
